+++ Source/General/Macro.vb
@@ -124,9 +124,9 @@ Public Class Macro
 
         If includeInteractive Then
             ret.Add(New Macro("$browse_file$", "Browse For File", GetType(String), "Filepath returned from a file browser."))
-            ret.Add(New Macro("$enter_text$", "Enter Text", GetType(String), "Text entered in a input box."))
-            ret.Add(New Macro("$enter_text:prompt$", "Enter Text (Params)", GetType(String), "Text entered in a input box."))
-            ret.Add(New Macro("$select:param1;param2;...$", "Select", GetType(String), "String selected from dropdown, to show a optional message the first parameter has to start with msg: and to give the items optional captions use caption|value. Example: $select:msg:hello;caption1|value1;caption2|value2$"))
+            ret.Add(New Macro("$enter_text$", "Enter Text", GetType(String), "Text entered in an input box."))
+            ret.Add(New Macro("$enter_text:prompt$", "Enter Text (Params)", GetType(String), "Text entered in an input box."))
+            ret.Add(New Macro("$select:param1;param2;...$", "Select", GetType(String), "String selected from dropdown, to show an optional message the first parameter has to start with msg: and to give the items optional captions use caption|value. Example: $select:msg:hello;caption1|value1;caption2|value2$"))
         End If
 
         If includeParam Then